moved RelayStubProvider from Util to DevTools
