+++ src/Apps/Artwork/Components/__stories__/ArtworkSidebar.story.tsx
@@ -14,7 +14,7 @@ import {
   RegularNonEditionedArtwork,
   VideoArtwork,
 } from "Apps/__test__/Fixtures/Artworks"
-import { RelayStubProvider } from "Utils/RelayStubProvider"
+import { RelayStubProvider } from "DevTools/RelayStubProvider"
 
 storiesOf("Styleguide/Artwork", module)
   .addDecorator(story => <RelayStubProvider>{story()}</RelayStubProvider>)
