+++ src/Apps/Artwork/Components/__stories__/ArtworkSidebar/ArtworkSidebarArtists.story.tsx
@@ -4,10 +4,10 @@ import {
   SingleNonFollowedArtist,
 } from "Apps/__test__/Fixtures/Artwork/Sidebar/Artists"
 import { ArtworkSidebarArtists as Artists } from "Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarArtists"
+import { RelayStubProvider } from "DevTools/RelayStubProvider"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Styleguide/Utils/Section"
-import { RelayStubProvider } from "Utils/RelayStubProvider"
 
 storiesOf("Styleguide/Artwork/Sidebar", module)
   .addDecorator(story => <RelayStubProvider>{story()}</RelayStubProvider>)
