+++ src/Apps/Artwork/Components/__stories__/ArtworkSidebar/ArtworkSidebarCommercial.story.tsx
@@ -5,10 +5,10 @@ import {
   FoSaleArtworkWithOneEdition,
 } from "Apps/__test__/Fixtures/Artwork/Sidebar/Commercial"
 import { ArtworkSidebarCommercial as Commercial } from "Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarCommercial"
+import { RelayStubProvider } from "DevTools/RelayStubProvider"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Styleguide/Utils/Section"
-import { RelayStubProvider } from "Utils/RelayStubProvider"
 
 storiesOf("Styleguide/Artwork/Sidebar", module)
   .addDecorator(story => <RelayStubProvider>{story()}</RelayStubProvider>)
