+++ src/Apps/Artwork/Components/__stories__/ArtworkSidebar/ArtworkSidebarMetadata.story.tsx
@@ -7,10 +7,10 @@ import {
   FilledOutMetadataOneEditionSet,
 } from "Apps/__test__/Fixtures/Artwork/Sidebar/ArtworkMetadata"
 import { ArtworkSidebarMetadata as Metadata } from "Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarMetadata"
+import { RelayStubProvider } from "DevTools/RelayStubProvider"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Styleguide/Utils/Section"
-import { RelayStubProvider } from "Utils/RelayStubProvider"
 
 storiesOf("Styleguide/Artwork/Sidebar", module)
   .addDecorator(story => <RelayStubProvider>{story()}</RelayStubProvider>)
