+++ src/Apps/Artwork/Components/__stories__/ArtworkSidebarLive.story.tsx
@@ -3,7 +3,7 @@ import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Styleguide/Utils/Section"
 
-import { RelayStubProvider } from "Utils/RelayStubProvider"
+import { RelayStubProvider } from "DevTools/RelayStubProvider"
 
 storiesOf("Styleguide/Artwork", module)
   .addDecorator(story => <RelayStubProvider>{story()}</RelayStubProvider>)
